requirements.txt
streamlit==1.33.0
tensorflow==2.15.0
numpy
Pillow
matplotlib
gdown

